Create Armstrong_Number.py
What is Armstrong number? - In number theory, a narcissistic number (also known as a pluperfect digital invariant (PPDI), an Armstrong number (after Michael F. Armstrong) or a plus perfect number) in a given number base. is a number that is the sum of its own digits each raised to the power of the number of digits.

- In these lines of code we will search for Armstrong numbers in the range we want.